paddleocr/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
paddlepaddle==2.6.2
paddleocr==2.9.1
numpy==1.26.4
Pillow==10.4.0
python-multipart==0.0.12
